+++ R/check.R
@@ -87,7 +87,7 @@ convert_and_check_my_args <- function(options) {
       no("timeout") || (length(timeout) == 1 && !is.na(timeout)),
       no("wd") || is_string(wd),
       no("fail_on_status") || is_flag(fail_on_status),
-      is_string(package) || is_flag(package),
+      is.null(package) || is_string(package) || is_flag(package),
       is_string(arch)
     )
   )