+++ R/eval.R
@@ -99,7 +99,10 @@
 #'   means no timeout.
 #' @param package Whether to keep the environment of `func` when passing
 #'   it to the other package. Possible values are:
-#'   * `FALSE`: reset the environment to `.GlobalEnv`. This is the default.
+#'   * `NULL` (the default): equivalent to `TRUE` if `func` inherits from
+#'     `"crate"` (i.e. was created with `carrier::crate()`), and `FALSE`
+#'     otherwise.
+#'   * `FALSE`: reset the environment to `.GlobalEnv`.
 #'   * `TRUE`: keep the environment as is.
 #'   * `pkg`: set the environment to the `pkg` package namespace.
 #' @param arch Architecture to use in the child process, for multi-arch
@@ -187,7 +190,7 @@ r <- function(
   user_profile = "project",
   env = rcmd_safe_env(),
   timeout = Inf,
-  package = FALSE,
+  package = NULL,
   arch = "same",
   ...
 ) {